+++ NEWS.md
@@ -1,7 +1,8 @@
 # dynamAedes 3.0.0
 * Fixes bug on dybamAedes.m at the regional scale when the species is Aedes aegypti.
 * Updates adci function
-* Implement Multiple introduction events through the multipleIntro function 
+* Implements Multiple introduction events through the multipleIntro function
+* Adds the get_rate function to compute mosquito temperature/photoperiod dependent rates
 
 # dynamAedes 2.2.9
 Substitutes legend.pos with legend.position in vignettes.